\name{cdists-package}
\alias{cdists-package}
\alias{cdists}
\docType{package}
\title{
\packageTitle{cdists}
}
\description{
\packageDescription{cdists}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cdists}
\packageIndices{cdists}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cdists}

Maintainer: \packageMaintainer{cdists}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
