\name{cdists-package}
\alias{cdists-package}
\alias{cdists}
\docType{package}
\title{
\packageTitle{cdists}
}
\description{
\packageDescription{cdists}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cdists}
\author{
\packageAuthor{cdists}

Maintainer: \packageMaintainer{cdists}
}
\references{
~~ Literature or other references for background information ~~
}